test push/some fixes
changes were made ti the gamemanager so its not looking for a heal or ammo thing anymore.

Added: i added a diffrent way to win you need to find the keys and go to truck instead of killing all the enemys
red truck has been added along with a muzzel flash pack and a level prefabs. im testing to see if i can push the level prefabs rn